+++ src/main/java/advent/of/code/year2019/intcode/IOProvider.java
@@ -1,9 +1,7 @@
 package advent.of.code.year2019.intcode;
 
-import java.math.BigInteger;
-
 public interface IOProvider {
-    BigInteger nextInt();
+    Long nextInt();
 
-    void putInt(BigInteger value);
+    void putInt(Long value);
 }
