+++ src/main/java/advent/of/code/year2019/intcode/operations/InputOperation.java
@@ -1,9 +1,7 @@
 package advent.of.code.year2019.intcode.operations;
 
-import advent.of.code.year2019.intcode.IntCodeProcess;
 import advent.of.code.year2019.intcode.IntCodeParameter;
-
-import java.math.BigInteger;
+import advent.of.code.year2019.intcode.IntCodeProcess;
 
 public class InputOperation implements Operation {
     @Override
@@ -18,7 +16,7 @@ public class InputOperation implements Operation {
 
     @Override
     public void execute(IntCodeProcess intCodeProcess, IntCodeParameter[] parameters) {
-        BigInteger[] program = intCodeProcess.getProgram();
+        Long[] program = intCodeProcess.getProgram();
         program[parameters[0].getPointer()] = intCodeProcess.getIoProvider().nextInt();
     }
 }
